wpt_babel/content-security-policy/worker-src/service-none.https.sub.html

<!DOCTYPE html>
<script src="/resources/testharness.js">
</script>
<script src="/resources/testharnessreport.js">
</script>
<script src="../support/testharness-helper.js">
</script>
<meta content="worker-src 'none'" http-equiv="Content-Security-Policy"/>
<script>
 "use strict";

var url = new URL("../support/ping.js", document.baseURI).toString();
assert_service_worker_is_blocked(url, "Same-origin service worker blocked by 'none'.");
</script>
